Add WikidataOverviewStats component
